+++ example/lib/routing/routes.dart
@@ -2,4 +2,5 @@ class AppRoutes {
   static const String splash = '/';
   static const String scanning = '/scanning';
   static const String newOtaUpdate = '/new_ota_Update';
+  static const String otaLogs = '/ota_logs';
 }